# docs/swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Sample API
  description: Demo REST API generated offline from Spring MVC annotations
  contact:
    name: API Team
    email: api@example.com
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  version: 1.0.0
servers:
- url: https://api.example.com/sample-api/
  description: Production
- url: https://staging.example.com/sample-api/
  description: Staging
- url: http://localhost:8080/sample-api/
  description: Local development
security:
- bearerAuth: []
...
paths:
  /api/v1/orders/{id}:
    get:
      tags:
      - orders
      summary: Get order by ID
      operationId: getOrder
      parameters:
      - name: id
        in: path
        description: Order ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Order found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
    delete:
      tags:
      - orders
      summary: Cancel an order
      operationId: cancelOrder
      parameters:
      - name: id
        in: path
        description: ID of the order to cancel
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Order cancelled
        "404":
          description: Order not found
  /api/v1/orders:
    get:
      tags:
      - orders
      summary: List all orders
      operationId: listOrders
      responses:
        "200":
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderDto'
    post:
      tags:
      - orders
      summary: Place a new order
      description: Creates a new order for the given customer.
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderDto'
        required: true
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDto'
...
  /api/v1/products/{id}:
    get:
      tags:
      - products
      summary: Get by ID
      description: Returns a single resource identified by its primary key.
      operationId: getById_1
      parameters:
      - name: id
        in: path
        description: Primary key of the resource
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Resource found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
        "404":
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
    delete:
      tags:
      - products
      summary: Delete by ID
      description: Permanently removes a resource.
      operationId: deleteById
      parameters:
      - name: id
        in: path
        description: Primary key of the resource to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "204":
          description: Resource deleted
        "404":
          description: Resource not found
  /api/v1/products:
    get:
      tags:
      - products
      summary: List all
      description: Returns every resource (unfiltered).
      operationId: listAll
      responses:
        "200":
          description: Full list returned
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
    post:
      tags:
      - products
      summary: Create product
      description: Adds a new product to the catalogue.
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDto'
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductDto'
  /api/v1/products/category/{category}:
    get:
      tags:
      - products
      summary: List products by category
      description: Returns all products belonging to the given category.
      operationId: listByCategory
      parameters:
      - name: category
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDto'
components:
  schemas:
    OrderDto:
      required:
      - customerId
      - total
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          format: int64
          readOnly: true
          example: 1
        customerId:
          type: integer
          description: ID of the customer who placed the order
          format: int64
          example: 42
        status:
          type: string
          description: Current order status
          example: PENDING
          enum:
          - PENDING
          - CONFIRMED
          - SHIPPED
          - DELIVERED
          - CANCELLED
        total:
          type: number
          description: Total order amount in EUR
          example: 149.99
        createdAt:
          type: string
          description: Timestamp when the order was placed
          format: date-time
          readOnly: true
      description: Order data transfer object
...
    ProductDto:
      required:
      - name
      - price
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier
          format: int64
          readOnly: true
          example: 1
        name:
          type: string
          description: Product name
          example: Wireless Keyboard
        description:
          type: string
          description: Detailed description
          example: Compact bluetooth keyboard
        price:
          type: number
          description: Unit price in EUR
          example: 49.99
        category:
          type: string
          description: Product category
          example: ELECTRONICS
          enum:
          - ELECTRONICS
          - CLOTHING
          - FOOD
          - BOOKS
        stock:
          type: integer
          description: Available stock quantity
          format: int32
          example: 150
      description: Product data transfer object
  securitySchemes:
    bearerAuth:
      type: http
      description: "JWT token obtained from the auth service (Authorization: Bearer\
        \ <token>)"
      scheme: bearer
      bearerFormat: JWT